Create nlp-bert_0.83844LBScore.ipynb
Real or Not? NLP with Disaster Tweets
Predict which Tweets are about real disasters and which ones are not

Twitter has become an important communication channel in times of emergency.
The ubiquitousness of smartphones enables people to announce an emergency they’re observing in real-time. Because of this, more agencies are interested in programatically monitoring Twitter (i.e. disaster relief organizations and news agencies).

But, it’s not always clear whether a person’s words are actually announcing a disaster.